+++ crates/symbolicator/src/services/symcaches/mod.rs
@@ -8,8 +8,7 @@ use anyhow::Error;
 use futures::future::BoxFuture;
 use sentry::{configure_scope, Hub, SentryFutureExt};
 use symbolic::common::{Arch, ByteView};
-use symbolic::debuginfo::Object;
-use symbolic::symcache::{self, SymCache, SymCacheWriter};
+use symbolic::symcache::{self, SymCache, SymCacheConverter};
 use thiserror::Error;
 
 use crate::cache::{Cache, CacheStatus};
@@ -28,6 +27,7 @@ use crate::utils::sentry::ConfigureScope;
 
 use self::markers::{SecondarySymCacheSources, SymCacheMarkers};
 
+use super::il2cpp::Il2cppService;
 use super::shared_cache::SharedCacheService;
 
 mod markers;
@@ -63,10 +63,10 @@ pub enum SymCacheError {
     Fetching(#[source] ObjectError),
 
     #[error("failed to parse symcache")]
-    Parsing(#[source] symcache::SymCacheError),
+    Parsing(#[source] symcache::Error),
 
     #[error("failed to write symcache")]
-    Writing(#[source] symcache::SymCacheError),
+    Writing(#[source] symcache::Error),
 
     #[error("malformed symcache file")]
     Malformed,
@@ -76,6 +76,9 @@ pub enum SymCacheError {
 
     #[error("failed to handle auxiliary BCSymbolMap file")]
     BcSymbolMapError(#[source] Error),
+
+    #[error("failed to handle auxiliary il2cpp line mapping file")]
+    Il2cppError(#[source] Error),
 
     #[error("symcache building took too long")]
     Timeout,
@@ -89,6 +92,7 @@ pub struct SymCacheActor {
     symcaches: Arc<Cacher<FetchSymCacheInternal>>,
     objects: ObjectsActor,
     bitcode_svc: BitcodeService,
+    il2cpp_svc: Il2cppService,
     threadpool: tokio::runtime::Handle,
 }
 
@@ -98,12 +102,14 @@ impl SymCacheActor {
         shared_cache_svc: Arc<SharedCacheService>,
         objects: ObjectsActor,
         bitcode_svc: BitcodeService,
+        il2cpp_svc: Il2cppService,
         threadpool: tokio::runtime::Handle,
     ) -> Self {
         SymCacheActor {
             symcaches: Arc::new(Cacher::new(cache, shared_cache_svc)),
             objects,
             bitcode_svc,
+            il2cpp_svc,
             threadpool,
         }
     }
@@ -322,20 +328,44 @@ impl SymCacheActor {
             Some(handle) => {
                 // TODO: while there is some caching *internally* in the bitcode_svc, the *complete*
                 // fetch request is not cached
-                let bcsymbolmap_handle = match handle.object_id().debug_id {
-                    Some(debug_id) => {
-                        self.bitcode_svc
-                            .fetch_bcsymbolmap(
-                                debug_id,
-                                handle.scope().clone(),
-                                request.sources.clone(),
-                            )
-                            .await
+                let fetch_bcsymbolmap = async {
+                    match handle.object_id().debug_id {
+                        Some(debug_id) => {
+                            self.bitcode_svc
+                                .fetch_bcsymbolmap(
+                                    debug_id,
+                                    handle.scope().clone(),
+                                    request.sources.clone(),
+                                )
+                                .await
+                        }
+                        None => None,
                     }
-                    None => None,
                 };
 
-                let secondary_sources = SecondarySymCacheSources { bcsymbolmap_handle };
+                let fetch_il2cpp = async {
+                    match handle.object_id().debug_id {
+                        Some(debug_id) => {
+                            tracing::trace!("Fetching line mapping");
+                            self.il2cpp_svc
+                                .fetch_line_mapping(
+                                    debug_id,
+                                    handle.scope().clone(),
+                                    request.sources.clone(),
+                                )
+                                .await
+                        }
+                        None => None,
+                    }
+                };
+
+                let (bcsymbolmap_handle, il2cpp_handle) =
+                    futures::future::join(fetch_bcsymbolmap, fetch_il2cpp).await;
+
+                let secondary_sources = SecondarySymCacheSources {
+                    bcsymbolmap_handle,
+                    il2cpp_handle,
+                };
 
                 self.symcaches
                     .compute_memoized(FetchSymCacheInternal {
@@ -373,15 +403,15 @@ fn write_symcache(
         object_handle.to_scope(scope);
     });
 
-    let mut symbolic_object = object_handle
+    let symbolic_object = object_handle
         .parse()
         .map_err(SymCacheError::ObjectParsing)?
         .unwrap();
 
     let markers = SymCacheMarkers::from_sources(&secondary_sources);
 
-    if let Object::MachO(ref mut macho) = symbolic_object {
-        if let Some(ref handle) = secondary_sources.bcsymbolmap_handle {
+    let bcsymbolmap_transformer = match secondary_sources.bcsymbolmap_handle {
+        Some(ref handle) => {
             let bcsymbolmap = handle
                 .bc_symbol_map()
                 .map_err(SymCacheError::BcSymbolMapError)?;
@@ -390,17 +420,43 @@ fn write_symcache(
                 handle.uuid,
                 object_handle
             );
-            #[allow(deprecated)]
-            macho.load_symbolmap(bcsymbolmap);
+            Some(bcsymbolmap)
         }
-    }
-
-    let file = File::create(&path)?;
-    let mut writer = BufWriter::new(file);
+        None => None,
+    };
+    let linemapping_transformer = match secondary_sources.il2cpp_handle {
+        Some(handle) => {
+            let il2cpp = handle.line_mapping().map_err(SymCacheError::Il2cppError)?;
+            tracing::debug!(
+                "Adding il2cpp line mapping {} to object {}",
+                handle.debug_id,
+                object_handle
+            );
+            Some(il2cpp)
+        }
+        None => None,
+    };
 
     tracing::debug!("Converting symcache for {}", object_handle.cache_key());
 
-    SymCacheWriter::write_object(&symbolic_object, &mut writer).map_err(SymCacheError::Writing)?;
+    let mut converter = SymCacheConverter::new();
+
+    if let Some(bcsymbolmap) = bcsymbolmap_transformer {
+        converter.add_transformer(bcsymbolmap);
+    }
+    if let Some(linemapping) = linemapping_transformer {
+        converter.add_transformer(linemapping);
+    }
+
+    converter
+        .process_object(&symbolic_object)
+        .map_err(SymCacheError::Writing)?;
+
+    let file = File::create(&path)?;
+    let mut writer = BufWriter::new(file);
+    converter
+        .serialize(&mut writer)
+        .map_err(SymCacheError::Io)?;
 
     let mut file = writer.into_inner().map_err(io::Error::from)?;
 
@@ -434,17 +490,17 @@ mod tests {
         let mut cache_config = CacheConfigs::default();
         cache_config.downloaded.retry_misses_after = Some(timeout);
 
-        let config = Arc::new(Config {
+        let config = Config {
             cache_dir: Some(cache_dir),
             connect_to_reserved_ips: true,
             caches: cache_config,
             ..Default::default()
-        });
+        };
 
         let cpu_pool = tokio::runtime::Handle::current();
         let caches = Caches::from_config(&config).unwrap();
         caches.clear_tmp(&config).unwrap();
-        let downloader = DownloadService::new(config);
+        let downloader = DownloadService::new(&config);
         let shared_cache = Arc::new(SharedCacheService::new(None).await);
         let objects = ObjectsActor::new(
             caches.object_meta,
@@ -452,9 +508,17 @@ mod tests {
             shared_cache.clone(),
             downloader.clone(),
         );
-        let bitcode = BitcodeService::new(caches.auxdifs, shared_cache.clone(), downloader);
+        let bitcode = BitcodeService::new(caches.auxdifs, shared_cache.clone(), downloader.clone());
+        let il2cpp = Il2cppService::new(caches.il2cpp, shared_cache.clone(), downloader);
 
-        SymCacheActor::new(caches.symcaches, shared_cache, objects, bitcode, cpu_pool)
+        SymCacheActor::new(
+            caches.symcaches,
+            shared_cache,
+            objects,
+            bitcode,
+            il2cpp,
+            cpu_pool,
+        )
     }
 
     /// Tests that a symcache is regenerated when it was created without a BcSymbolMap
@@ -518,9 +582,12 @@ mod tests {
         // symcache will be obfuscated.
         let symcache_file = symcache_actor.fetch(fetch_symcache.clone()).await.unwrap();
         let symcache = symcache_file.parse().unwrap().unwrap();
-        let line_info = symcache.lookup(0x5a75).unwrap().next().unwrap().unwrap();
-        assert_eq!(line_info.filename(), "__hidden#42_");
-        assert_eq!(line_info.symbol(), "__hidden#0_");
+        let sl = symcache.lookup(0x5a75).next().unwrap();
+        assert_eq!(
+            sl.file().unwrap().full_path(),
+            "__hidden#41_/__hidden#41_/__hidden#42_"
+        );
+        assert_eq!(sl.function().name(), "__hidden#0_");
 
         // Copy the plist and bcsymbolmap to the temporary symbol directory so that the SymCacheActor can find them.
         fs::copy(
@@ -540,9 +607,12 @@ mod tests {
         // be obfuscated like before.
         let symcache_file = symcache_actor.fetch(fetch_symcache.clone()).await.unwrap();
         let symcache = symcache_file.parse().unwrap().unwrap();
-        let line_info = symcache.lookup(0x5a75).unwrap().next().unwrap().unwrap();
-        assert_eq!(line_info.filename(), "__hidden#42_");
-        assert_eq!(line_info.symbol(), "__hidden#0_");
+        let sl = symcache.lookup(0x5a75).next().unwrap();
+        assert_eq!(
+            sl.file().unwrap().full_path(),
+            "__hidden#41_/__hidden#41_/__hidden#42_"
+        );
+        assert_eq!(sl.function().name(), "__hidden#0_");
 
         // Sleep long enough for the negative cache entry to become invalid.
         std::thread::sleep(TIMEOUT);
@@ -551,8 +621,11 @@ mod tests {
         // symcache are unobfuscated.
         let symcache_file = symcache_actor.fetch(fetch_symcache.clone()).await.unwrap();
         let symcache = symcache_file.parse().unwrap().unwrap();
-        let line_info = symcache.lookup(0x5a75).unwrap().next().unwrap().unwrap();
-        assert_eq!(line_info.filename(), "Sources/Sentry/SentryMessage.m");
-        assert_eq!(line_info.symbol(), "-[SentryMessage initWithFormatted:]");
+        let sl = symcache.lookup(0x5a75).next().unwrap();
+        assert_eq!(
+            sl.file().unwrap().full_path(),
+            "/Users/philipphofmann/git-repos/sentry-cocoa/Sources/Sentry/SentryMessage.m"
+        );
+        assert_eq!(sl.function().name(), "-[SentryMessage initWithFormatted:]");
     }
 }